Move home dir logic to package environment
