+++ pkg/helm/environment/default_home_dir.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package helmpath
+package environment
 
 import (
 	"github.com/casimir/xdg-go"
