+++ pkg/helm/environment/default_home_dir_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helmpath
+package environment
 
 import (
 	"runtime"
